+++ model.validata.com/Order/OrderViewModel.cs
@@ -9,7 +9,7 @@ namespace model.validata.com.Order
         public int OrderId { get; set; }
         public DateTime OrderDate { get; set; }
         [Required]
-        public float TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
 
         [Required]
         public int ProductCount { get; set; }
